Fixed a bug with creating a profile
