+++ CoursePlatform/accounts/apps.py
@@ -3,6 +3,3 @@ from django.apps import AppConfig
 
 class AccountsConfig(AppConfig):
     name = 'accounts'
-
-    def ready(self):
-        import accounts.signals
